requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
isort==5.13.2
magic-filter==1.0.12
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
pyee==11.1.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schedule==1.2.1
soupsieve==2.5
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.2
yarl==1.9.4
